initial DE analysis (don't think this is super great right now)
Tried cNMF in R. Unfortunately, I can't run it in the terminal since my version of python is 3.12 and cNMF does not work with that version of python.
Will remove later.

Trying to run cNMF in Jupyter notebook. code to create the matrix, tables of barcodes, and genes is included here